FIX：统一固定宽度整数类型风格
WHY：项目中 unsigned/unsigned long/long long 等原生整数类型与 uint32_t/int64_t 等固定宽度类型混用，ESP32-C6 工具链下 uint32_t 实际映射为 unsigned long，机械统一类型后容易触发 printf/sscanf 格式不匹配并被 -Werror 阻断编译。

WHAT：将业务代码中的显式 unsigned/long 风格类型和转换统一为 uint*_t/int*_t；将日志、Web JSON、Shell 输出、屏幕显示和 OTA 解析中的格式串迁移到 PRI*/SCN* 宏；保留 C++ 用户自定义字面量要求的 unsigned long long 标准签名；修复黑匣子快照 flags 初始化 warning 并移除未使用函数。

TEST：

编译OK

手动测试OK

冒烟测试OK

+++ components/app/blackbox_service/src/blackbox_service.cpp
@@ -70,7 +70,7 @@ esp_err_t set_snapshot_interval_s(uint32_t seconds, const char* source) {
         return err;
     }
     append_text_event("blackbox: config source=%s snapshot_interval_s=%lu", source == nullptr ? "unknown" : source,
-                      static_cast<unsigned long>(seconds));
+                      static_cast<uint32_t>(seconds));
     return ESP_OK;
 }
