+++ components/app/screen/src/pages/battery_page.cpp
@@ -8,6 +8,7 @@
 #include "pages/battery_page.h"
 
 #include <algorithm>
+#include <cinttypes>
 #include <cmath>
 #include <cstdio>
 
@@ -105,10 +106,10 @@ void format_fixed_digits(char* line, size_t line_size, double value, const char*
  * @param total_seconds 总秒数。
  */
 void format_duration(char* line, size_t line_size, const char* prefix, uint64_t total_seconds) {
-    snprintf(line, line_size, "%s%02llu:%02llu:%02llu", prefix == nullptr ? "" : prefix,
-             static_cast<unsigned long long>(total_seconds / 3600),
-             static_cast<unsigned long long>((total_seconds / 60) % 60),
-             static_cast<unsigned long long>(total_seconds % 60));
+    snprintf(line, line_size, "%s%02" PRIu64 ":%02" PRIu64 ":%02" PRIu64, prefix == nullptr ? "" : prefix,
+             static_cast<uint64_t>(total_seconds / 3600),
+             static_cast<uint64_t>((total_seconds / 60) % 60),
+             static_cast<uint64_t>(total_seconds % 60));
 }
 
 } // namespace
